requirements.txt
fastapi
uvicorn
python-nmap
ping3
